Deprecated geoscales functions

These functions were renamed under the harmonized naming convention shared with the timescales package. The old names warn and forward to their replacements; they will be removed before the 1.0 release.

Usage

geo_recast(...)

geo_filter(...)

geo_prune(...)

geo_attach_geometry(x, geom, by = NULL, level = NULL, ...)

geo_area(...)

geo_levels(...)

geo_rank(...)

geo_weights(...)

geo_regions(...)

geo_family(...)

geo_nests(...)

geo_ancestry(...)

geo_children(...)

geo_parents(...)

geo_descendants(...)

geo_ancestors(...)

geo_share(...)

geo_geometry(...)

geo_layout(...)

geo_autoplot(...)

geo_plot(...)

geo_register_rule(...)

geo_get_rule(...)

geo_list_rules(...)

geo_clear_rules(...)

geo_register_provider(...)

geo_provider(...)

geo_list_providers(...)

geoscale_levels(...)

is_valid_level(...)

geoscale_from_leaves(leaves, levels, ...)

CORE_LEVELS

Arguments

...

Arguments forwarded to the replacement function.

Value

See the replacement function.

Details

Data operations and object transforms (verb_geoscale): geo_recast() -> recast_geoscale() (or the recast() generic), geo_filter() -> filter_geoscale(), geo_prune() -> prune_geoscale(), geo_attach_geometry() -> attach_geometry_geoscale(), geo_area() -> add_area_geoscale().

Properties and queries (geoscale_*): geo_levels() -> geoscale_geoframes(), geo_rank() -> geoscale_rank(), geo_weights() -> geoscale_weights(), geo_regions() -> geoscale_regions(), geo_family() -> geoscale_family(), geo_nests() -> geoscale_nests(), geo_ancestry() -> geoscale_ancestry(), geo_children() -> geoscale_children(), geo_parents() -> geoscale_parents(), geo_descendants() -> geoscale_descendants(), geo_ancestors() -> geoscale_ancestors(), geo_share() -> geoscale_share(), geo_geometry() -> geoscale_geometry(), geo_layout() -> geoscale_layout(), geo_autoplot() -> geoscale_autoplot(), geo_plot() -> geoscale_plot().

Registries: geo_register_rule() -> register_geo_rule(), geo_get_rule() -> get_geo_rule(), geo_list_rules() -> list_geo_rules(), geo_clear_rules() -> clear_geo_rules(), geo_register_provider() -> register_geo_provider(), geo_provider() -> get_geo_provider(), geo_list_providers() -> list_geo_providers().